refactor(mm-next): move header related component into `/header` folder
