+++ packages/mirror-media-next/components/header/normal/header.js
@@ -7,11 +7,11 @@ import {
   PROMOTION_LINKS,
   SOCIAL_MEDIA_LINKS,
   SUB_BRAND_LINKS,
-} from '../constants'
-import { DEFAULT_NORMAL_SECTIONS_DATA } from '../constants/header'
+} from '../../../constants'
+import { DEFAULT_NORMAL_SECTIONS_DATA } from '../../../constants/header'
 
-import SubscribeMagazineButton from '../components/subscribe-papermag-mb-btn'
-import { SEARCH_URL } from '../config/index.mjs'
+import SubscribeMagazineButton from './subscribe-papermag-mb-btn'
+import { SEARCH_URL } from '../../../config/index.mjs'
 import EntranceButton from './entrance-button'
 import Logo from './logo'
 import MemberLoginButton from './member-login-button'
@@ -19,13 +19,13 @@ import MobileSidebar from './mobile-sidebar'
 import NavSections from './nav-sections'
 import NavTopics from './nav-topics'
 import PromotionLinks from './promotion-links'
-import SearchBarDesktop from './search-bar-desktop'
-import SearchBarInput from './search-bar-input'
+import SearchBarDesktop from '../shared/search-bar-desktop'
+import SearchBarInput from '../shared/search-bar-input'
 import SubBrandList from './sub-brand-list'
 
-import { useDisplayAd } from '../hooks/useDisplayAd'
+import { useDisplayAd } from '../../../hooks/useDisplayAd'
 
-const GPTAd = dynamic(() => import('../components/ads/gpt/gpt-ad'), {
+const GPTAd = dynamic(() => import('../../ads/gpt/gpt-ad'), {
   ssr: false,
 })
