+++ packages/mirror-media-next/components/header/premium/premium-header.js
@@ -1,15 +1,15 @@
 import React, { useState, useRef, useEffect } from 'react'
 import styled from 'styled-components'
 
-import SearchBarDesktop from './search-bar-desktop'
-import { Z_INDEX } from '../constants'
-import { DEFAULT_PREMIUM_SECTIONS_DATA } from '../constants/header'
-import SearchBarInput from './search-bar-input'
-import Logo from './logo'
+import SearchBarDesktop from '../shared/search-bar-desktop'
+import { Z_INDEX } from '../../../constants'
+import { DEFAULT_PREMIUM_SECTIONS_DATA } from '../../../constants/header'
+import SearchBarInput from '../shared/search-bar-input'
+import Logo from '../normal/logo'
 import PremiumMobileSidebar from './premium-mobile-sidebar'
 import PremiumNavSections from './premium-nav-sections'
 import PremiumMemberLoginButton from './premium-member-login-button'
-import { SEARCH_URL } from '../config/index.mjs'
+import { SEARCH_URL } from '../../../config/index.mjs'
 
 /**
  * @typedef {import('./premium-mobile-sidebar').H2AndH3Block} H2AndH3Block
@@ -155,7 +155,7 @@ const MobilePremiumMemberLoginButton = styled(PremiumMemberLoginButton)`
 
 /**
  *
- * @typedef {import('../apollo/fragments/section').Section[]} PremiumHeaderSections
+ * @typedef {import('../../../apollo/fragments/section').Section[]} PremiumHeaderSections
  *
  * @typedef PremiumHeaderData
  * @property {PremiumHeaderSections} sections
