+++ packages/mirror-media-next/components/header/premium/premium-nav-sections.js
@@ -3,7 +3,7 @@
 //TODO: Replace <a> to <Link> for Single Page Application
 import styled, { css } from 'styled-components'
 /**
- * @typedef {import('../type/theme').Theme} Theme
+ * @typedef {import('../../../type/theme').Theme} Theme
  */
 
 const colorCss = css`
