+++ packages/mirror-media-next/components/header/shared/search-bar-input.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components'
-import { maxWidth } from '../styles/media'
+import { maxWidth } from '../../../styles/media'
 
 const SearchInputWrapper = styled.div`
   font-size: 16px;
