+++ packages/mirror-media-next/components/story/premium/index.js
@@ -15,7 +15,7 @@ import DonateLink from '../shared/donate-link'
 import SubscribeLink from '../shared/subscribe-link'
 import ArticleMask from '../shared/article-mask'
 import { useMembership } from '../../../context/membership'
-import ShareHeader from '../../shared/share-header'
+import ShareHeader from '../../header/share-header'
 import Footer from '../../shared/footer'
 import { useDisplayAd } from '../../../hooks/useDisplayAd'
 import { Z_INDEX } from '../../../constants/index'
